% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_dataset.R
\name{create_test_dataset}
\alias{create_test_dataset}
\title{Create test dataset on specific station for prediction}
\usage{
create_test_dataset(station_number)
}
\arguments{
\item{station_number}{number that means the number of each station.(1 ~ 144)}
}
\value{
a dataset containing feature and rental count data on 'station_number' station, 2015
}
\description{
A function to create test dataset on 'station_number' bike station by preprocessing bike rental history and weather data in 2015.
}
\examples{
test_dataset <- create_test_dataset(1)

}
