% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_train_model.R
\name{create_train_model}
\alias{create_train_model}
\title{Create random-forest training model for bicycle rental prediction.}
\usage{
create_train_model(train_dataset)
}
\arguments{
\item{train_dataset}{Training dataset created by create_train_dataset()}
}
\value{
random forest training model
}
\description{
Create random-forest training model for bicycle rental prediction.
}
\examples{
\dontrun{train_dataset <- create_train_dataset(3)
rf_model <- create_train_model(train_dataset)}
}
