% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestData.R
\name{createTestData}
\alias{createTestData}
\title{Create Train Data for prediction In StationNum.}
\usage{
createTestData(stationNum)
}
\arguments{
\item{stationNum}{Station Number (1 ~ 144)}
}
\description{
Create Train Data for prediction In StationNum.
}
\examples{
createTrainData(1)
}
