% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainData.R
\name{createTrainData}
\alias{createTrainData}
\title{Create Train Data for prediction In StationNum.}
\usage{
createTrainData(stationNum)
}
\arguments{
\item{stationNum}{Station Number (1 ~ 144)}
}
\description{
Create Train Data for prediction In StationNum.
}
\examples{
createTrainData(1)
}
