% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_bicycle_rental.R
\name{daily_bicycle_rental}
\alias{daily_bicycle_rental}
\title{Visualize amount of bicycle rental at each day of week.}
\usage{
daily_bicycle_rental()
}
\description{
A function analyzing bicycle rental pattern on each day of week and visualizing analyzed result.
}
\examples{
\dontrun{daily_bicycle_rental()}
}
