% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_bike_rental.R
\name{daily_bike_rental}
\alias{daily_bike_rental}
\title{Visualize amount of bicycle rental at each day of week.}
\usage{
daily_bike_rental()
}
\description{
A function analyzing bike rental pattern on each day of week and visualizing analyzed result.
}
\examples{
\dontrun{daily_bike_rental()}
}
