% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_train_model.R
\name{extract_features}
\alias{extract_features}
\title{Extract feature columns from train/test dataset}
\usage{
extract_features(data)
}
\arguments{
\item{data}{data with feature columns and others}
}
\value{
data containing only feature columns
}
\description{
Extract feature columns from train/test dataset
}
