% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDemandOfBikesUsingRF.R
\name{extractFeatures}
\alias{extractFeatures}
\title{extract features in test data frame.}
\usage{
extractFeatures(data)
}
\arguments{
\item{data}{testing data frame for prediction. Testing data frame is created by conducting createTestData()}
}
\description{
extract feature data from test Data.
feature : season, month, hour, day of week, temperature, humidity, rainfall, affect of festival.
}
