% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{festivalData}
\alias{festivalData}
\title{festival data In Daejeon from 2013 to 2015}
\format{A data frame with 116 rows and 5 variables:
\describe{
  \item{nearStation}{tashu station near the festival location}
  \item{startDate}{datetime when festival start}
  \item{endDate}{datetime when festival end}
}}
\usage{
festivalData
}
\description{
A dataset containing festival data in Daejeon, from 2013 to 2015
}
\keyword{datasets}
