% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAmountOfRentalEachDayOfWeek.R
\name{getAmountOfRentalEachDayOfWeek}
\alias{getAmountOfRentalEachDayOfWeek}
\title{Visualize amount of bicycle rental at each day of week.}
\usage{
getAmountOfRentalEachDayOfWeek()
}
\description{
Visualize amount of bicycle rental at each day of week.
}
\examples{
\dontrun{
getAmountOfRentalEachDayOfWeek()
}
}
