% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAmountOfRentalEachMonth.R
\name{getAmountOfRentalEachMonth}
\alias{getAmountOfRentalEachMonth}
\title{Get monthly amount of rental for 3 years.}
\usage{
getAmountOfRentalEachMonth()
}
\description{
Get monthly amount of rental for 3 years.
}
\examples{
getAmountOfRentalEachMonth()
}
