% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeavilyUsedStations.R
\name{getHeavilyUsedStations}
\alias{getHeavilyUsedStations}
\title{Get top 10 stations that have the most amount of rentals from 2013 to 2015.}
\usage{
getHeavilyUsedStations()
}
\value{
The visualized image
}
\description{
Get top 10 stations that have the most amount of rentals from 2013 to 2015.
}
\examples{
getHeavilyUsedStations()
}
