% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMostUsedPath.R
\name{getMostUsedPath}
\alias{getMostUsedPath}
\title{Get most Used 10 Pathes from 2013 to 2015 And visualize the result on map.}
\usage{
getMostUsedPath()
}
\description{
Get most Used 10 Pathes from 2013 to 2015 And visualize the result on map.
}
\examples{
getMostUsedPath()
}
