% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTop10Pathes.R
\name{getTop10Pathes}
\alias{getTop10Pathes}
\title{Get Top 10 Pathes that were most used from 2013 to 2015 And visualize the result on map.}
\usage{
getTop10Pathes()
}
\value{
data frame that contains most used top 10 pathes from 2013 to 2015
}
\description{
Get Top 10 Pathes that were most used from 2013 to 2015 And visualize the result on map.
}
\examples{
top10PathesData <- getTop10Pathes()
}
