% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTop10Stations.R
\name{getTop10Stations}
\alias{getTop10Stations}
\title{Visualize top 10 stations that were most used from 2013 to 2015.}
\usage{
getTop10Stations()
}
\value{
Data frame that contains top 10 most used stations from 2013 to 2015
}
\description{
Draw a plot that visualized most used top 10 stations on barchart.
}
\examples{
\dontrun{getTop10Stations()}
}
