% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsageByTemperature.R
\name{getUsageByTemperature}
\alias{getUsageByTemperature}
\title{Show the change of rental amount according to Temperature.}
\usage{
getUsageByTemperature()
}
\value{
A graph that present usage change by temperature and month
}
\description{
Show the change of rental amount according to Temperature.
}
\examples{
getUsageByTemperature()
}
