% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_bicycle_rental.R
\name{monthly_bicycle_rental}
\alias{monthly_bicycle_rental}
\title{Visualize the change of bicycle rental amount by temperature and each month.}
\usage{
monthly_bicycle_rental()
}
\description{
A function drawing a plot that shows change of temperature and bicycle rental ratio in each month.
}
\examples{
\dontrun{monthly_bicycle_rental()}
}
