% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_bike_rental.R
\name{monthly_bike_rental}
\alias{monthly_bike_rental}
\title{Visualize the change of bicycle rental amount by temperature and each month.}
\usage{
monthly_bike_rental()
}
\description{
A function drawing a plot that shows change of temperature and bike rental ratio in each month.
}
\examples{
monthly_bike_rental()
}
