% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bicycle_rental.R
\name{predict_bicycle_rental}
\alias{predict_bicycle_rental}
\title{Predict hourly Demand of bicycle in 2015.}
\usage{
predict_bicycle_rental(rf_model, test_dataset)
}
\arguments{
\item{rf_model}{random forest prediction model create by create_train_model()}

\item{test_dataset}{testing dataset}
}
\value{
test_dataset with predictive result.
}
\description{
predict hourly amount of bicycle rental in 2015 using random forest algorithm. Create prediction model using 'train_dataset' and forecast demand of bicycle rental according to the condition of 'test_dataset'
}
\examples{
\dontrun{train_dataset <- create_train_dataset(3)
test_dataset <- create_test_dataset(3)
rf_model <- create_train_model(train_dataset)
test_dataset <- predict_bicycle_rental(rf_model, test_dataset)}

}
