% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tashu2013}
\alias{tashu2013}
\title{Bike rental history on Tashu in 2013.}
\format{A data frame with 1033827 rows and 4 variables:
\describe{
  \item{RENT_STATION}{station Number where bike was rented.}
  \item{RENT_DATE}{date time when bike was rented. yyyy-mm-dd hh:mm:ss}
  \item{RETURN_STATION}{station Number where bike was returned.}
  \item{RETURN_DATE}{date time when bike was returned. yyyy-mm-dd hh:mm:ss}
}}
\source{
\url{https://www.data.go.kr/dataset/3068538/fileData.do/}
}
\usage{
tashu2013
}
\description{
A dataset containing bike rental history(rent station, rent date, return station, return date)
}
\keyword{datasets}
