% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tashuStationData}
\alias{tashuStationData}
\title{Bike station data on Tashu.}
\format{A data frame with 144 rows and 9 variables:
\describe{
  \item{NUM}{station Number}
  \item{NumOFBIKE_RACK}{the number of bike rack}
  \item{GEODATA_lat}{latitude of bike station}
  \item{GEODATA_lon}{longitude of bike station}
}}
\usage{
tashuStationData
}
\description{
A dataset containing 144 bike station
}
\keyword{datasets}
