% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10_pathes.R
\name{top10_pathes}
\alias{top10_pathes}
\title{Visualize Top 10 Pathes that were most used from 2013 to 2015.}
\usage{
top10_pathes()
}
\description{
Visualize Top 10 Pathes that were most used from 2013 to 2015.
}
\examples{
\dontrun{top10_pathes()}

}
