% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10_paths.R
\name{top10_paths}
\alias{top10_paths}
\title{Visualize Top 10 Pathes that were most used from 2013 to 2015.}
\usage{
top10_paths()
}
\description{
Visualize Top 10 Pathes that were most used from 2013 to 2015.
}
\examples{
\dontrun{top10_paths()}

}
