% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10_stations.R
\name{top10_stations}
\alias{top10_stations}
\title{Visualize top 10 stations that were most used from 2013 to 2015.}
\usage{
top10_stations()
}
\value{
Data frame that contains top 10 most used stations from 2013 to 2015
}
\description{
Draw a plot that visualized most used top 10 stations on barchart.
}
\examples{
top10Stations()
}
