% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestData.R
\name{createTestData}
\alias{createTestData}
\title{Create test data frame for prediction.}
\usage{
createTestData(stationNum)
}
\arguments{
\item{stationNum}{Station Number (1 ~ 144)}
}
\description{
This function create test data frame for prediction.
}
\examples{
createTrainData(1)
}
