% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestData.R
\name{createTestData}
\alias{createTestData}
\title{Create testing data frame for prediction.}
\usage{
createTestData(stationNum)
}
\arguments{
\item{stationNum}{number that means the number of each station.(1 ~ 144)}
}
\value{
a data frame that contains test Data in 'stationNum' station, 2015
columns : datetime, season, rentMonth, rentHour, rentWeekday, temperature, humidity, rainfall, isFestival, RrentCount(Real number of rental), PrentCount(NA, Predictive number of rental would be filled)
}
\description{
Create testing data frame in 'stationNum' bike station by preprocessing bike rental history, weather and festival data in 2015.
}
\examples{
\dontrun{
testData <- createTestData(1)
}

}
