% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestData.R
\name{createTestData}
\alias{createTestData}
\title{Create test data frame for prediction.}
\usage{
createTestData(stationNum)
}
\arguments{
\item{stationNum}{Station Number (1 ~ 144)}
}
\value{
a data frame that contains hourly (datetiem, season, month, hour, day of week, temperature, humidity, rainfall, isFestival, real number of rental in stationNum station, )
}
\description{
Create test data frame for prediction.
}
\examples{
createTrainData(1)
}
