% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainData.R
\name{createTrainData}
\alias{createTrainData}
\title{Create training Data for prediction In Station stationNum.}
\usage{
createTrainData(stationNum)
}
\arguments{
\item{stationNum}{Station Number (1 ~ 144)}
}
\description{
This is a function that create training data for forecasting bike demand in station "stationNum".
}
\examples{
createTrainData(1)
}
