% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainData.R
\name{createTrainData}
\alias{createTrainData}
\title{Create training Data for prediction In Station stationNum.}
\usage{
createTrainData(stationNum)
}
\arguments{
\item{stationNum}{number that means the number of each station.(1 ~ 144)}
}
\value{
a data frame that present hourly amount of bike rental in 'stationNum' station from 2013 to 2014.
columns: datetime, season, month, hourm, day of week, temperature, humidity, rainfall, festival, rental count
}
\description{
Create training data frame in 'stationNum' bike station by preprocessing bike rental history, weather and festival data from 2013 to 2014.
}
\examples{
trainData <- createTrainData(1)
}
