% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_train_model.R, R/predict_bike_rental.R
\name{extract_features}
\alias{extract_features}
\title{Create random-forest training model for bike rental prediction.}
\usage{
extract_features(data)

extract_features(data)
}
\arguments{
\item{train_dataset}{Training dataset created by create_train_dataset()}

\item{train_dataset}{training data for creating prediction model}

\item{test_dataset}{testing data.}

\item{numOftree}{number of tree in random Forest}
}
\value{
random forest training model

test_dataset with predictive result.
}
\description{
predict hourly amount of bike rental in 2015 using random forest algorithm. Create prediction model using 'train_dataset' and forecast demand of bike rental according to the condition of 'test_dataset'
}
\examples{
train_dataset <- create_train_dataset(3)
rf_model <- create_train_model(train_dataset)
train_dataset <- createtrain_dataset(3)
test_dataset <- createtest_dataset(3)
predictDemandOfBikesUsingRF(train_dataset, test_dataset, numOftree = 50)

}
