% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{extractFeatures}
\alias{extractFeatures}
\title{Feature extraction Function}
\usage{
extractFeatures(data)
}
\arguments{
\item{data}{#정류소 별 대여 데이터(+계절, 월, 시간, 요일, 기온, 습도, 강수량, 축제 여부)}
}
\description{
This function extracts feature data that affects demand of bikes in Tashu.
}
\examples{
extractFeatures()
}
\keyword{extraction}
\keyword{feature}
