% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{extractFeatures}
\alias{extractFeatures}
\title{extract features in test data frame.}
\usage{
extractFeatures(data)
}
\arguments{
\item{data}{#정류소 별 대여 데이터(+계절, 월, 시간, 요일, 기온, 습도, 강수량, 축제 여부)}
}
\description{
This function extracts features that affect demand of bike rental in test data frame.
feature : season, month, hour, day of week, temperature, humidity, rainfall, affect of festival.
}
\examples{
extractFeatures()
}
