% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAmountOfRentalEachDayOfWeek.R
\name{getAmountOfRentalEachDayOfWeek}
\alias{getAmountOfRentalEachDayOfWeek}
\title{Get monthly amount of rental Each day of week}
\usage{
getAmountOfRentalEachDayOfWeek()
}
\description{
Get monthly amount of rental Each day of week
}
\examples{
getAmountOfRentalEachDayOfWeek()
}
