% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAmountOfRentalEachMonth.R
\name{getAmountOfRentalEachMonth}
\alias{getAmountOfRentalEachMonth}
\title{Visualize amount of bicycle rental at each month.}
\usage{
getAmountOfRentalEachMonth()
}
\description{
Draw a plot that visualize monthly average amount of bike rental from 2013 to 2015.
}
\examples{
\dontrun{getAmountOfRentalEachMonth()}
}
