% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTop10Pathes.R
\name{getTop10Pathes}
\alias{getTop10Pathes}
\title{Visualize Top 10 Pathes that were most used from 2013 to 2015.}
\usage{
getTop10Pathes()
}
\description{
Visualize Top 10 Pathes that were most used from 2013 to 2015.
}
\examples{
\dontrun{getTop10Pathes()}
}
