% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTop10Stations.R
\name{getTop10Stations}
\alias{getTop10Stations}
\title{Get top 10 stations that have the most amount of rentals from 2013 to 2015.}
\usage{
getTop10Stations()
}
\description{
Draw two plots that visualized most used top 10 stations on barchart and map.
}
\examples{
getTop10Stations()
}
