% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsageByTemperature.R
\name{getUsageByTemperature}
\alias{getUsageByTemperature}
\title{Get usage pattern By temperature and month.}
\usage{
getUsageByTemperature()
}
\value{
visualized image that present usage change by temperature and month
}
\description{
Get usage pattern By temperature and month.
}
\examples{
getUsageByTemperature()
}
