% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsageByTemperature.R
\name{getUsageByTemperature}
\alias{getUsageByTemperature}
\title{Show the change of amount of bike rental by Temperature.}
\usage{
getUsageByTemperature()
}
\description{
This function draw a plot that show change of average temperature and average bike rental ratio in each month.
}
\examples{
getUsageByTemperature()
}
