% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDemandOfBikesUsingRF.R
\name{predictDemandOfBikesUsingRF}
\alias{predictDemandOfBikesUsingRF}
\title{Predict hourly Demand of Bike in 2015.}
\usage{
predictDemandOfBikesUsingRF(stationNum, isImportance, numOfNtree, type)
}
\arguments{
\item{stationNum}{A number from 1 to 144}

\item{isImportance}{if TRUE, visualize the ranking of Feature importance.}

\item{numOfNtree}{random Forest Ntree number}

\item{type}{if 0, classification. if 1, regression}
}
\description{
Predict hourly Demand of Bike in 2015.
}
\examples{
predictDemandOfBikesUsingRF(3, TRUE, 50, 1)
}
