% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDemandOfBikesUsingRF.R
\name{predictDemandOfBikesUsingRF}
\alias{predictDemandOfBikesUsingRF}
\title{Predict hourly Demand of Bike in 2015.}
\usage{
predictDemandOfBikesUsingRF(trainData, testData, numOftree = 50)
}
\arguments{
\item{trainData}{training data for creating prediction model}

\item{testData}{testing data.}

\item{numOftree}{number of tree in random Forest}
}
\value{
testData with predictive result.
}
\description{
predict hourly amount of bike rental in 2015 using random forest algorithm. Create prediction model using 'trainData' and forecast demand of bike rental according to the condition of 'testData'
}
\examples{
trainData <- createTrainData(3)
testData <- createTestData(3)
predictDemandOfBikesUsingRF(trainData, testData, numOftree = 50)
}
