% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDemandOfBikesUsingRF.R
\name{predictDemandOfBikesUsingRF}
\alias{predictDemandOfBikesUsingRF}
\title{Predict hourly Demand of Bike in 2015.}
\usage{
predictDemandOfBikesUsingRF(stationNum, isImportance, numOftree, type)
}
\arguments{
\item{isImportance}{if TRUE, visualize the ranking of Feature importance.}

\item{numOftree}{number of tree in random Forest}

\item{type}{if 0, random Forest for classification. if 1, random Forest for regression}

\item{trainData}{training data for creating prediction model}

\item{testData}{testing data for prediction}
}
\description{
Predict hourly Demand of Bike in 2015.
}
\examples{
trainData <- createTrainData(3)
testData <- createTestData(3)
predictDemandOfBikesUsingRF(trainData, testData, TRUE, 50, 1)
}
